Move ‘session’ module under ‘linnote.core’
