+++ linnote/core/utils/__init__.py
@@ -7,3 +7,5 @@ Miscellaneaous tools for the project.
 Author: Anatole Hanniet, 2016-2018.
 License: Mozilla Public License, see 'LICENSE.txt' for details.
 """
+
+from .session import configure_session, session
